+++ modules/AWSPowerShell/AWSAliases.ps1
@@ -2268,9 +2268,13 @@ Set-Alias -Name CE-ListCostAllocationTags -Value Get-CECostAllocationTagList
 Set-Alias -Name List-CECostCategoryDefinitions -Value Get-CECostCategoryDefinitionList
 Set-Alias -Name List-CECostCategoryDefinitionList -Value Get-CECostCategoryDefinitionList
 Set-Alias -Name CE-ListCostCategoryDefinitions -Value Get-CECostCategoryDefinitionList
+Set-Alias -Name List-CESavingsPlansPurchaseRecommendationGeneration -Value Get-CESavingsPlansPurchaseRecommendationGenerationList
+Set-Alias -Name List-CESavingsPlansPurchaseRecommendationGenerationList -Value Get-CESavingsPlansPurchaseRecommendationGenerationList
+Set-Alias -Name CE-ListSavingsPlansPurchaseRecommendationGeneration -Value Get-CESavingsPlansPurchaseRecommendationGenerationList
 Set-Alias -Name CE-ListTagsForResource -Value Get-CEResourceTag
 Set-Alias -Name Provide-CEAnomalyFeedback -Value Set-CEAnomalyFeedback
 Set-Alias -Name CE-ProvideAnomalyFeedback -Value Set-CEAnomalyFeedback
+Set-Alias -Name CE-StartSavingsPlansPurchaseRecommendationGeneration -Value Start-CESavingsPlansPurchaseRecommendationGeneration
 Set-Alias -Name CE-TagResource -Value Add-CEResourceTag
 Set-Alias -Name CE-UntagResource -Value Remove-CEResourceTag
 Set-Alias -Name CE-UpdateAnomalyMonitor -Value Update-CEAnomalyMonitor