+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -5914,6 +5914,13 @@ $CE_Completers = {
             break
         }
 
+        # Amazon.CostExplorer.GenerationStatus
+        "Get-CESavingsPlansPurchaseRecommendationGenerationList/GenerationStatus"
+        {
+            $v = "FAILED","PROCESSING","SUCCEEDED"
+            break
+        }
+
         # Amazon.CostExplorer.Granularity
         {
             ($_ -eq "Get-CECostAndUsage/Granularity") -Or
@@ -6043,6 +6050,7 @@ $CE_map = @{
     "Dimension"=@("Get-CEDimensionValue")
     "Feedback"=@("Get-CEAnomaly","Set-CEAnomalyFeedback")
     "Frequency"=@("Update-CEAnomalySubscription")
+    "GenerationStatus"=@("Get-CESavingsPlansPurchaseRecommendationGenerationList")
     "Granularity"=@("Get-CECostAndUsage","Get-CECostAndUsageWithResource","Get-CECostForecast","Get-CEReservationCoverage","Get-CEReservationUtilization","Get-CESavingsPlansCoverage","Get-CESavingsPlansUtilization","Get-CEUsageForecast")
     "LookbackPeriodInDays"=@("Get-CEReservationPurchaseRecommendation","Get-CESavingsPlansPurchaseRecommendation")
     "Metric"=@("Get-CECostForecast","Get-CEUsageForecast")
@@ -6134,8 +6142,10 @@ $CE_SelectMap = @{
                "Get-CEUsageForecast",
                "Get-CECostAllocationTagList",
                "Get-CECostCategoryDefinitionList",
+               "Get-CESavingsPlansPurchaseRecommendationGenerationList",
                "Get-CEResourceTag",
                "Set-CEAnomalyFeedback",
+               "Start-CESavingsPlansPurchaseRecommendationGeneration",
                "Add-CEResourceTag",
                "Remove-CEResourceTag",
                "Update-CEAnomalyMonitor",
@@ -54167,7 +54177,7 @@ $TRS_Completers = {
         # Amazon.TranscribeService.CLMLanguageCode
         "New-TRSLanguageModel/LanguageCode"
         {
-            $v = "en-AU","en-GB","en-US","es-US","hi-IN"
+            $v = "de-DE","en-AU","en-GB","en-US","es-US","hi-IN","ja-JP"
             break
         }