+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2AwsNetworkPerformanceMetricSubscription-Cmdlet.cs
@@ -28,7 +28,7 @@ using Amazon.EC2.Model;
 namespace Amazon.PowerShell.Cmdlets.EC2
 {
     /// <summary>
-    /// Describes the curent Infrastructure Performance metric subscriptions.<br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
+    /// Describes the current Infrastructure Performance metric subscriptions.<br/><br/>This cmdlet automatically pages all available results to the pipeline - parameters related to iteration are only needed if you want to manually control the paginated output. To disable autopagination, use -NoAutoIteration.
     /// </summary>
     [Cmdlet("Get", "EC2AwsNetworkPerformanceMetricSubscription")]
     [OutputType("Amazon.EC2.Model.Subscription")]